fix: move mq code into database
